refactor: folder names
