+++ kotlin/code/playground/src/main/kotlin/coroutines/StructuredConcurrency.kt
@@ -1,6 +1,6 @@
 @file:Suppress("Unused", "TooManyFunctions", "MagicNumber")
 
-package threads
+package coroutines
 
 import io.github.oshai.KotlinLogging
 import kotlinx.coroutines.Deferred
