+++ kotlin/code/playground/src/main/kotlin/functional/Raises.kt
@@ -1,6 +1,6 @@
 @file:Suppress("Unused", "MagicNumber", "UnusedPrivateProperty")
 
-package func
+package functional
 
 import arrow.core.Either
 import arrow.core.NonEmptyList
