+++ kotlin/code/playground/src/main/kotlin/language/ContextReceivers.kt
@@ -1,6 +1,6 @@
 @file:Suppress("Unused", "MagicNumber")
 
-package lang
+package language
 
 import kotlinx.coroutines.runBlocking
