+++ kotlin/code/playground/src/main/kotlin/language/Operators.kt
@@ -1,6 +1,6 @@
 @file:Suppress("Unused", "MagicNumber", "UnusedPrivateProperty")
 
-package lang
+package language
 
 import arrow.core.getOrElse
 import arrow.core.raise.Raise
